refactor: refactor model for passengers form
