+++ src/app/booking/models/forms.model.ts
@@ -1,9 +1,9 @@
 import { AgeStatus } from 'src/app/core/models/age-status.model';
 import { ContactForm } from './contact-form.model';
-import { PassengersForm } from './passengers-form.model';
+import { PassengerData } from 'src/app/core/models/passengers.model';
 
 export type FormsData = {
-  passengersFrom?: PassengersForm[] | null;
+  passengersFrom?: PassengerData[] | null;
   contactForm?: ContactForm | null;
 };
 
@@ -11,6 +11,6 @@ export type PassengersFormData = {
   id: string;
   cardHead: string;
   ageStatus: AgeStatus;
-  addPassengersForm: (props: PassengersForm) => void;
+  addPassengersForm: (props: PassengerData) => void;
   setFormFullField: (props: { id: string; value: boolean }) => void;
 };
