--- src/app/booking/models/passengers-form.model.ts
@@ -1,8 +1,0 @@
-export type PassengersForm = {
-  firstName: string;
-  lastName: string;
-  gender: string;
-  dateOfBird: string;
-  baggage?: string;
-  baggageCount?: string;
-};
